added unit tests for logout

+++ server/src/main/java/service/HTTP/Logout.java
@@ -10,6 +10,10 @@ public class Logout extends HTTPService {
     public Response run(Request req) throws DataAccessException {
         AuthRequest r = (AuthRequest) req;
 
+        if (r.getAuthorization() == null) {
+            throw new DataAccessException("Error: unauthorized", 401);
+        }
+
         int data = authDAO.delete(r.getAuthorization());
         if (data == 0) {
             throw new DataAccessException("Error: unauthorized", 401);
